+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.9.0] - TODO
+## [0.9.0] - 2023-02-28
 ### Fixes
 - Added a limiter to `FP::to_vec_general` and downstream string formatting to prevent easy resource
   exhaustion problems. Note that `max_ufp` is set to 4096 for default formatters.
@@ -12,7 +12,7 @@
 - Added IEEE-754 related items for `FP`
 - Added fixed point support to `ExtAwi::from_str` to quickly leverage `ExtAwi::from_str_general` 
 
-## [0.8.0] - 2022-01-17
+## [0.8.0] - 2023-01-17
 ### Crate
 - MSRV 1.66